refactor(models): remote duplicate Contract model, keep OtcState model
